refactor: Update headers in RegistryClient
This commit refactors the code in the `RegistryClient` module by updating the headers used in the HTTP request. The `ACCEPT` header is now conditionally set based on the value of the `options.full_metadata` flag. If the flag is true, the header is set to "application/json", otherwise it is set to "application/vnd.npm.install-v1+json; q=1.0, application/json; q=0.8, */*".